Virtual Travel agent(python)/requirements.txt
Flask==2.3.3
requests==2.31.0
python-dotenv==1.0.0
PyYAML==6.0.1
reportlab==4.0.7
pandas==2.1.3
Pillow==10.1.0
